execute bbb tests only if feature has been enabled
